requirements.txt
asgiref
certifi
chardet
defusedxml
dj-database-url
Django
django-allauth
django-ckeditor
django-js-asset
django-multiselectfield
gunicorn
idna
oauthlib
Pillow
psycopg2
psycopg2-binary
python3-openid
pytz==2020.1
requests
requests-oauthlib
sqlparse
urllib3
whitenoise
